refactor: split VolunteerOpportunitiesList god file into focused modules
frontend/src/components/VolunteerOpportunitiesList.tsx had grown to 1,757
lines mixing an icon library, a calendar widget, dropdown primitives, and
data-fetching alongside the list rendering itself (#866). Split into a
VolunteerOpportunitiesList/ folder: icons.tsx, MiniCalendar.tsx,
FilterDropdown.tsx, OpportunityListItem.tsx, OpportunityResultsList.tsx,
useVolunteerOpportunitiesData.ts, useCitySuggestions.ts, and a
VolunteerOpportunitiesList.tsx orchestrator - mirroring the existing
Header/ folder convention. No behavior change.
